Some tweaks after running on a new Pop!_os install
No longer install Regolith
Prompt for DVD/Bluray tools
Prompt for NodeJS
Fix some package names

+++ ubuntu/setup-desktop.sh
@@ -1,24 +1,24 @@
 #!/usr/bin/env bash
 sudo add-apt-repository universe
 
-sudo apt install -y audacity libbluray-bdj libaacs0 libdvd-pkg \
-darktable evolution evolution-ews ffmpeg ffmpeg-devel \
+sudo apt install -y audacity  \
+darktable evolution evolution-ews ffmpeg libavcodec-dev \
 flameshot fonts-firacode gimp gnome-tweaks kdiff3 obs-studio xclip \
 libavcodec-extra ubuntu-restricted-extras gnome-shell-extension-impatience \
 gnome-shell-extension-no-annoyance 
 
 # bluray/dvd
-sudo dpkg-reconfigure libdvd-pkg
-mkdir -p ~/.config/aacs/
-curl -L -o ~/.config/aacs/KEYDB.cfg http://vlc-bluray.whoknowsmy.name/files/KEYDB.cfg
+read -p "Install Bluray/DVD tools? " -n 1 -r
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    sudo apt install libbluray-bdj libaacs0 libdvd-pkg
+    sudo dpkg-reconfigure libdvd-pkg
+    mkdir -p ~/.config/aacs/
+    curl -L -o ~/.config/aacs/KEYDB.cfg http://vlc-bluray.whoknowsmy.name/files/KEYDB.cfg
+fi
 
-# keybase
-# keybasepkg=keybase_amd64.deb
-# curl -O https://prerelease.keybase.io/$keybasepkg
-# sudo dpkg -i $keybasepkg
-# sudo apt-get install -f -y
-# run_keybase
-# rm $keybasepkg
+
 
 # slack
 curl -o /tmp/slack-desktop-4.2.0-amd64.deb  https://downloads.slack-edge.com/linux_releases/slack-desktop-4.2.0-amd64.deb
@@ -27,15 +27,12 @@ sudo apt install -fy
 
 
 
-# Install dash to dock extension for Pop!_OS
+# Install Ubuntu dock extension for Pop!_OS
 if grep -Fq "Pop" /etc/os-release
 then
-    sudo apt install gnome-shell-extension-dashtodock
+    sudo apt install gnome-shell-extension-ubuntu-dock
 fi
 
-# Install Regolith Desktop - https://regolith-linux.org
-sudo add-apt-repository -y ppa:regolith-linux/release
-sudo apt install -y regolith-desktop
 
 # Install cascadia code font
